+++ src/Core/Client/Adapter/Guzzle3.php
@@ -31,7 +31,7 @@ class Guzzle3 extends Configurable implements AdapterInterface
      */
     public function execute($request, $endpoint)
     {
-        $baseUri = $request->getIsServerRequest() ? $endpoint->getServerUri() : $endpoint->getCoreBaseUri();
+        $baseUri = $request->getIsServerRequest() ? $endpoint->getServerUri() : $endpoint->getBaseUri();
         $uri = $baseUri.$request->getUri();
         $guzzleRequest = $this->getGuzzleClient()->createRequest(
             $request->getMethod(),